+++ index.js
@@ -69,7 +69,7 @@ function initializeControllers() {
     const { startInterval } = require("./middlewares/mongoConnector/helper.js");
     startInterval();
     const { currentDirectory } = require(`./common-storage/common-${process.env.STORAGE_TYPE}.js`);
-    const { preCompanySetup, } = require("./modules/Company/controller2.js");
+    const { preCompanySetup, } = require("./modules/company/controller2.js");
     app.get("/api/v1/setPresetCompany/:id", (req, res) => {
         if (req.params && req.params.id && req.params.id === config.PRECOMPANYKEY) {
             preCompanySetup();
@@ -86,7 +86,7 @@ function initializeControllers() {
     require('./modules/sprints/init.js').init(app);
     require('./modules/logTime/init.js').init(app);
     require('./modules/milestone/init.js').init(app);
-    require('./modules/Company/init.js').init(app);
+    require('./modules/company/init.js').init(app);
     require('./modules/trackerDownload/init.js').init(app);
     require('./modules/notification/notification-middleware/init').init(app);
     require('./modules/notification/prepare-notification-data/init').init(app);
@@ -110,7 +110,7 @@ function initializeControllers() {
     require('./modules/Project/init').init(app);
     require('./modules/Teams/init').init(app);
     require('./modules/tours/init').init(app);
-    require('./modules/AdvanceGlobalFilter/init.js').init(app);
+    require('./modules/advance-global-filter/init.js').init(app);
     require('./modules/settings/settingCurrency/init').init(app);
     require('./modules/settings/settingNotifications/init').init(app);
     require('./modules/projectRules/init').init(app);
