+++ modules/check-install-step/createCompany.js
@@ -14,7 +14,7 @@ const mainCtr = require("./controller.js");
 const installStepsFilePath = __dirname + "/../../installationSteps.json";
 const defaultSubscriptionDataRef = require("./defaultSubscriptionData.js")
 const serviceFun = require("../serviceFunction.js");
-const { updateCompanyFun } = require("../Company/controller/updateCompany.js");
+const { updateCompanyFun } = require("../company/controller/updateCompany.js");
 const { updateUserFun, getUserByQueyFun } = require("../usersModule/controller.js");
 const { storeRefferalCode } = require("../affiliate/controller.js");
 const { handleCreateCompanyDataStorageFun } = require(`../../common-storage/common-${process.env.STORAGE_TYPE}.js`);
