+++ modules/notification/email-notification-handler/controllerV2.js
@@ -1,7 +1,7 @@
 const { MongoDbCrudOpration} = require("../../../utils/mongo-handler/mongoQueries");
 const { SCHEMA_TYPE } = require("../../../Config/schemaType")
 const logger = require("../../../Config/loggerConfig")
-const { getCompanyDataFun } = require("../../Company/controller/updateCompany");
+const { getCompanyDataFun } = require("../../company/controller/updateCompany");
 exports.fetchDetailsOfCompanies = (companies = []) => {
   return new Promise((resolve, reject) => {
     try {
