+++ modules/notification/notification-middleware/controllerV2.js
@@ -1,7 +1,7 @@
 const pushController = require('./push-controllerV2')
 const emailController = require('./email-controllerV2')
 const { NODE_ENV } = require("../../../Config/config");
-const { getCompanyDataFun } = require("../../Company/controller/updateCompany");
+const { getCompanyDataFun } = require("../../company/controller/updateCompany");
 const socketEmitter = require('../../../event/socketEventEmitter');
 
 const logger = require('../../../Config/loggerConfig');
