+++ modules/tasks/helpers/task_class_Mongo.js
@@ -16,7 +16,7 @@ const socketEmitter = require('../../../event/socketEventEmitter');
 const { addCommentCollection, updateCommentCollection } = require('../../comments/controller')
 const { updateMainChat } = require('../../MainChats/controller');
 const { replaceObjectKey } = require("../../auth/helper");
-const { emitListener } = require("../../Company/eventController.js");
+const { emitListener } = require("../../company/eventController.js");
 const { createCustomFields } = require("./helper.js");
 const { removeCache } = require('../../../utils/commonFunctions.js');
 const { updateRemainingTime } = require('../../logTime/controllerV2.js');
